Additional threading work.
It appears to be safe to remove many of the synchronized blocks given
that almost all of our properties are atomic. This should speed up the
library and make the code cleaner and easier to follow. The one part
where we absolutely need synchronization is during cleanup of the
MBBaseRequest object, where before there was likely a potential bug
that could cause networkActivityStopped to be called twice for a
request.