Make cancellations more thread-safe.
There was a small chance that if one thread was in the middle of main()
while someone else attempted to cancel the operation on a different
thread, that a crash could occur. Change how cancellation works so that
cancellation should be much safer. Now, if anyone calls cancel from a
different thread, just mark that the operation should be cancelled and
wait until the next iteration of the operation runloop occurs to do the
actual cancellation.