Mark runloop as running before CFRunLoopRun for proper cleanup.
The current code was calling [self setRunLoopIsRunning:YES] after
the call to CFRunLoopRun(), which meant the code in finish would
never actually call CFRunLoopStop().  This seems to have been fine
for iOS5 and earlier, but in iOS6, the thread would just hang in
the queue, eventually filling up the limit of outstanding requests
and blocking any further activity.