Fix potential race condition.
If a caller cancels a request and immediately releases it, it was
possible for the MBURLRequestOperation to be released by the operation
queue before it had a chance to clean up its own runloop. This change
ensures that the operation is cancelled properly and that the runloop
is cleaned up by performing this action on the main thread. The block
retains the operation until everything is complete.