Fix a potential race condition.
It was possible for the MBBaseRequest to be deallocated on the main
thread when the MBURLConnectionOperation was just about to send a
message to it. To fix this potential race condition, we now make the
MBURLConnectionOperation retain its delegate. This is safe as we are
very careful to always break this retain cycle when the operation is
finished.

The other option was to add lots of @synchronized blocks around every
access to the MBURLConnectionOperation's delegate. That code would be
slower and harder to follow.